impl `get()` of `storageEngine` and test it.

+++ include/utils.h
@@ -10,9 +10,9 @@ class Utils {
  public:
   static std::string GenRandString(const uint16_t);
   static std::string MakeFixedLength(const uint32_t, const int);
-  static uint64_t GenLocation(const uint32_t, const uint32_t);
-  static uint32_t LocationDecodeFileNum(const uint64_t);
-  static uint32_t LocationDecodeFileOffset(const uint64_t);
+  static uint64_t GenLocation(const uint8_t, const uint64_t);
+  static uint8_t LocationDecodeFileNum(const uint64_t);
+  static uint64_t LocationDecodeFileOffset(const uint64_t);
 };
 
 }  // namespace minikv
